Added salting
Yes, I do know that salting is used to prevent hackers from finding passwords and seeing that multiple people have the same password. In this case, it's just used so that the admin and user passwords can be simple and the encryption isn't easily findable through just looking through common words.